update scripts and config_lqx

+++ kernel-build/bak_config.sh
@@ -6,7 +6,11 @@ lqx_dir="/usr/src/linux-5.8.14-liquorix"
 
 echo "backing up lqx config..."
 if [ -d ${lqx_dir} ]; then
-   cat ${lqx_dir}/.config > ${bak_dir}/config_lqx
+   if [ -f "${lqx_dir}/.config" ]; then
+      cat ${lqx_dir}/.config > ${bak_dir}/config_lqx
+   else
+      echo "lqx config not found, skipping..."
+   fi
 else
    echo "lqx dir not found, skipping..."
 fi
